Fixing fxdate param no longer optional
This happened due to the new foler sync not generating default arg values. I don't mind losing default arg values (optional args) currently as I don't realy on them and it actually looks safer this was as it's now not possible to forget an important parameter when calling the procedure.