CALCULATECONTRACTMETRICS - minor performance optimization
Using a temp table for yltDerivedMeasures because the query optimizer
seems to make a bad decision when the result of this part of the CTE is
empty. The next link in the CTE chain does a group-by and the engine
probably tries to fold it into the previous step, which results in 5-7s
delay even if the result of yltDerivedMeasures is empty.