Support for LockedFxRate + remove unused cols
Unfortunately, I made changes related to two topics in the same commit instead of doing them one at a time.

The main topic of this commit is introducing the ability for a scenario to use boundfx for inforce portlayers. I've added two fields to the portlayers_scenario table which specify the date and rate of the portfolio in case the portlayer is inforce and the scenario uses the lockBoundFxRate flag. This fx rate is used by the calculate_metrics procedure, and is also passed into blocks (including ylt blocks) for optional use by powerbi.

The second topic of this commit is removing periodstart/stop columns from ylt blocks as they are not used and take up space and slow down import.