Fixing seasonal premium calculations
The seasonality table was incomplete. It was missing data for periods with no losses. When a period has no losses in any year, we use the pro-rata distribution of premium. When it has no losses in a current period but does in the rest of year, the seasonal factor is simply 0.

One important caveat is that we use the inforce layer to determine the share of layer duration, as leap would return a slightly different proportion than non-leap years. By itself, this would not be a big deal, but we have a "select distinct" which has "shareofayerduration" in the select list and that would return duplicate rows if we didn't filter for "inforce" layers only.