Another fix to seasonality
We can't just use periods for inforce layers, projections might get sliced up at different days of the year. Instead, to avoid having different shareoflayerduration for different layerviews of the same layerid (due to leap year), I group and use min_by to force always using non-leap year in this part of the calculation. Not sure if using just plain min or max would be more correct. Neither is 100% correct, but portlayers have leap years and yelt doesn't so it's always going to be an approximation.