BoundFX:
Added boundfx rate/date to all inforce portfolios, regardless of sceario (todo: we can move this earlier, to staging.portlayer).

This is done to allow powerbi to toggle using boundfxrate on/off regarless of what the scenario says about using boundfx. We want flexibility in PowerBI so users can explore effects of changes like this one.