Moving fx columns from portlayer_scenario to portlayer:
Since BoundFx Rate/Date do not depend on scenarios, we can move them to the staging schema.

Note that FX needs to be calculated before portlayers are populated because it is required to populate portlayer's boundfxrate.

Also note that I've added and started using a function for verifying that a value is not null: economic_model_management.verify_not_null. This is a validation tool I want to use to ensure I spot errors early. The primary use case is to ensure mandatory joins are successful. It should be used with left joins as inner joins just silently don't return a row when the join doesn't succeed.