FX - handled entierly in the DB (no longer in PowerBI)
1. FXRate is no longer used in PowerBI, all blocks now have converted USD
amounts using either the boundfx or the scenariofx. This makes DAX code
simpler and a fair bit quicker (in tandem with removing the dynamic
climate load factor manipulation).

2. fixed an issue with staging.fxrate table (was not seeing last date
due to using < instead of <= in the query that generates dates in
the POPULATE_FX() stored procedure.

3. removed verify_not_null because it was still throwing exceptions
non deterministically due to the else part being evaluated even when the
result was not coming from the else branch.