Removed resolve_scenario_data from proess_model
Resolution of scenario data in now done from the scenario editor, in chunks and in parallel.

When invoked from the DB we invoke it from load_all.

This allows running in parallel and in smaller chunks from the scenario editor (e.g. only resolve retrocontracts if that's the table that changed, and resolve different scenario tables in parallel).