Reducing granularity of subjectylt
We're never reading ylt data at the portlayer level in PowerBI, so
I'm reducing the granularity to the minimial necessary level. This
reduces the amount of data to be sent to PowerBI by about 70%.

If needed, we can implement lower-granularity-level calculation in
PowerBI, using the slow YELPT-based DAX approach. That way, we have the
results quickly at retro contract level, and if we only need the ylt
data for specific portlayers, we can get that too.